Simplifies unused ExceptionHandler
The ExceptionHandler currently isn't actually being used anywhere,
however it might still prove useful. If it does, it won't need to ignore
certain exceptions--something else will always be handling issues and
sending appropriate responses. So, it is still here and able to send and
log 500 responses. We'll see what happens.