Simplifies Listener back to String rawRequest
The Listener was getting too smart by parsing headers for content length
to determine if there was a body to grab. It should be so much simpler:
just keep listening and appending to a string until there isn't anything
left to collect.

The RequestBuilder therefore reverts to accepting a String instead of an
ArrayList. However, the details of parsing the rawRequest have changed a
bit, as it is now expecting a body component.