Do-While syntax in Listener
Was experiencing a weird error in browser--Listener's bufferedReader
wasn't recognizing itself as "ready", and consequently wasn't entering
the while ready loop to append each char of the request.

A hack solution was to read the first character outside of the loop,
then enter the loop to complete the process. That seemed dumb though.

By switching the syntax to Java's "do-while" style, the body evaluates
before the while expression, which is all the bufferedreader needs to
recognize itself as ready.